backend/requirements.txt
faiss_cpu==1.12.0
fastapi==0.122.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
langchain_core==1.1.0
langchain_groq==1.1.0
numpy==1.26.4
pandas==2.3.3
pydantic==2.12.5
python-dotenv==1.2.1
pyvi==0.1.1
scikit_learn==1.7.2
selenium==4.38.0
sentence_transformers==4.1.0
transformers==4.46.3
torch==2.7.1
tqdm==4.67.1
uvicorn==0.23.2